+++ src/test/java/org/embulk/output/s3/TestS3FileOutputPlugin.java
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package org.embulk.output;
+package org.embulk.output.s3;
 
 public class TestS3FileOutputPlugin
 {
